backend/requirements.txt
streamlit
streamlit-extras
pandas
requests
pyperclip
openpyxl
azure-storage-blob
python-dotenv
autogen-agentchat
autogen-ext[azure]
autogen-ext[openai]
psycopg2-binary
azure-keyvault-secrets
azure-functions
azure-functions-durable
openai
autogen-core
requests_oauthlib
azure-search-documents